Use Optim version of trust region
