+++ REQUIRE
@@ -4,9 +4,9 @@ DataFrames
 Distributions
 DocOpt
 Dtree
-ForwardDiff 0.1.8 0.2
+ForwardDiff 0.2
 JLD
-Optim 0.5 0.5.1
+Optim 0.5+
 FITSIO 0.8.4
 WCS 0.1.3
 Grid
